refactor: introduce packages/shared for byte-identical API/UI code
Adds a new kuroshiro-shared workspace package holding the code that was
verbatim-duplicated between the API and UI: the maintenance result types,
the device-model/firmware sync result types, the plugin Data Source Mode
constant/types, and the screen-shell rendering helpers. Both packages
import these from 'kuroshiro-shared' instead of keeping their own copy,
closing the two cross-package clone groups fallow flagged on 2026-08-22
(see docs/adr/0020).

Closes #846
