+++ .fallow-baselines/dupes.baseline.json
@@ -1,10 +1,8 @@
 {
   "clone_groups": [
-    "packages/api/src/maintenance/maintenance.service.ts:11-54|packages/ui/src/types.ts:140-183",
     "packages/api/src/devices/display.service.ts:290-300|packages/api/src/devices/display.service.ts:84-94",
-    "packages/ui/src/stores/deviceModels.ts:59-69|packages/ui/src/stores/firmware.ts:52-62",
+    "packages/ui/src/stores/deviceModels.ts:60-70|packages/ui/src/stores/firmware.ts:53-63",
     "packages/api/src/plugins/plugins.service.ts:232-243|packages/api/src/plugins/plugins.service.ts:348-359",
-    "packages/api/src/device-models/screen-shell.ts:24-36|packages/ui/src/utils/screenShell.ts:20-28",
     "packages/api/src/devices/display.service.ts:521-533|packages/api/src/mashup/services/mashup-renderer.service.ts:60-71",
     "packages/api/src/screens/screens.service.ts:116-121|packages/api/src/screens/screens.service.ts:180-185",
     "packages/api/src/devices/display.service.ts:532-540|packages/api/src/mashup/services/mashup-renderer.service.ts:70-78",
@@ -13,11 +11,9 @@
     "packages/api/src/test/fixtures.ts:88-107|packages/api/src/test/mockDeviceModelsService.ts:5-24"
   ],
   "clone_fingerprints": [
-    "dup:d32ddf4a:2",
     "dup:04d1333c:2",
     "dup:b3dbc735:2",
     "dup:e767379f:2",
-    "dup:48b60dd6:2",
     "dup:0a6a4764:2",
     "dup:e5bd8803:2",
     "dup:3219d03d:2",
@@ -26,16 +22,14 @@
     "dup:5304ed0c:2"
   ],
   "normalized_clone_fingerprints": [
-    "dup:a3006799:2",
     "dup:c61b2099:2",
-    "dup:c77b3abb6f87acd9-5:2",
-    "dup:c77b3abb6f87acd9-1:2",
     "dup:c77b3abb6f87acd9-4:2",
-    "dup:c77b3abb6f87acd9-8:2",
-    "dup:f2fdd24f:2",
+    "dup:c77b3abb6f87acd9-1:2",
     "dup:c77b3abb6f87acd9-7:2",
-    "dup:c77b3abb6f87acd9-3:2",
+    "dup:f2fdd24f:2",
     "dup:c77b3abb6f87acd9-6:2",
+    "dup:c77b3abb6f87acd9-3:2",
+    "dup:c77b3abb6f87acd9-5:2",
     "dup:c77b3abb6f87acd9-2:2"
   ]
 }
