+++ AGENTS.md
@@ -14,6 +14,8 @@ Default vocabulary: `needs-triage`, `needs-info`, `ready-for-agent`, `ready-for-
 
 Single-context: `CONTEXT.md` + `docs/adr/` at the repo root. See `docs/agents/domain.md`.
 
+Code that is byte-identical in API and UI lives in `packages/shared`; see ADR-0020 for what does and doesn't belong there.
+
 ### Fallow (dead code, duplication, complexity)
 
 `pnpm fallow:ci` runs in CI against committed baselines; fixing a baselined finding means re-running `pnpm fallow:baseline` in the same PR. See `docs/agents/fallow.md`.
