+++ README.md
@@ -180,6 +180,7 @@ For local hacking or deployment inspiration, check out [`docker-compose.yml`](./
 
 - [`packages/api`](./packages/api) — The NestJS backend
 - [`packages/ui`](./packages/ui) — The Vue 3 + Vuetify frontend
+- [`packages/shared`](./packages/shared) — Code that is byte-identical in the API and UI
 
 ---
