+++ docs/adr/0020-shared-package-holds-only-identical-cross-package-code.md
@@ -1,0 +1,21 @@
+# `packages/shared` holds only code that is byte-identical in API and UI today
+
+A fallow scan on 2026-08-22 found two cross-package clone groups — the maintenance result types (`packages/api/src/maintenance/maintenance.service.ts` ↔ `packages/ui/src/types.ts`) and the screen shell (`packages/api/src/device-models/screen-shell.ts` ↔ `packages/ui/src/utils/screenShell.ts`, function-for-function identical) — plus ten more type names duplicated verbatim across the two packages, and a hand-copied `DataSourceMode` literal union in the UI that silently drifts from the API's `DATA_SOURCE_MODES` const whenever a mode is added or renamed. The API and UI are deliberately independent packages (separate `tsconfig`s, separate build pipelines, one TypeORM/entity-typed and one wire/JSON-typed), and until now there was nowhere to put code both need without picking a side. We introduce a third workspace package, `kuroshiro-shared` (`packages/shared`), to hold exactly that code — and only that code: symbols whose definition is identical, today, in both packages. `Device`, `Screen`, `DeviceModel`, `Palette` and friends stay hand-written in the UI, because their wire shape (`Date` fields serialized to `string`) genuinely differs from the API's entity shape — moving them would force a fake shared type neither side actually uses as-is.
+
+`packages/shared` is source-only (a single `src/index.ts` barrel, no subpath exports, no build step) and is a `devDependency` of both `packages/api` and `packages/ui`, not a `dependency`: the production Docker stage runs `pnpm install --prod` against only the API's `package.json`, which would fail to resolve a `workspace:*` dependency it can't see. Instead, `packages/api/tsup.config.ts` sets `noExternal: ['kuroshiro-shared']` so the API's bundle inlines it, and Vite bundles it into the UI the same way it bundles any other workspace import. The package itself has no `"type": "module"` — the API is CommonJS-ish under `Node16` module resolution, and an ESM-typed source package makes `tsc` fail resolving value imports from it (TS1479).
+
+`build:migrations` compiles unbundled (`tsc`, no tsup), and the production image's `pnpm install --prod` never installs `kuroshiro-shared` at all, so a *value* import from shared inside a TypeORM entity or migration would become a `require('kuroshiro-shared')` that crashes `migration:run` in the image — a failure mode that wouldn't show up until deploy. This is enforced mechanically, not by convention: `packages/api/eslint.config.mjs` adds an `ts/no-restricted-imports` rule scoped to `src/**/*.entity.ts`, `src/migrations/**/*.ts`, and `src/config/typeorm.config.ts` that allows `import type { ... } from 'kuroshiro-shared'` but errors on a value import. `plugin-data-source.entity.ts` is the one entity that needs shared's `DataSourceMode`/`DataSourceLiteralValue` today, and it only ever needs them as types.
+
+The moved screen-shell functions take a new minimal structural type, `ScreenShellTarget` (`{ model: { cssClasses, cssVariables }, palette: { frameworkClass } }`), rather than either package's full `DeviceModel`/`Palette`. The API's `DeviceRenderTarget` and the UI's `RenderTarget` both satisfy it structurally with no casts, so neither package's richer render-target type needs to change shape. The two packages' screen-shell spec files — previously two copies of the same assertions plus a shared golden fixture in a repo-root `test/` directory — fold into one spec inside `packages/shared/src/__test__/`; the root `test/` directory and the API's `tsconfig.json` exclude working around it are both gone.
+
+## Considered Options
+
+- **Accept the duplication via `duplicates.ignoredClones`.** Rejected: the `DataSourceMode` case has already drifted once (the UI's union didn't get the `'literal'` mode added at the same time as the API's), which is exactly the failure mode a fallow baseline exists to catch, not to permanently silence.
+- **Make `packages/shared` the full wire contract, with a serialize layer translating entities to it.** Rejected for now: `Device`/`Screen`/`DeviceModel`/`Palette`'s `Date`-vs-`string` divergence is real and load-bearing (the API's entities need `Date` for TypeORM/business logic; the UI only ever sees the JSON-serialized `string`), and building a serialize layer to unify them is a much bigger change than today's actual duplication justifies. Worth revisiting if the wire-type surface grows enough to be worth the machinery.
+- **`pnpm deploy` in the production Docker stage**, so the prod image gets a real, installed copy of `kuroshiro-shared` instead of relying on bundling. Rejected for now: `packages/shared` carries no runtime code the entities or migrations need at runtime (the restricted-imports rule enforces exactly that), so there's nothing for a deployed copy to serve. Reconsider if that constraint ever gets lifted.
+
+## Consequences
+
+- `DeviceModelSyncResult.syncedAt` and `FirmwareSyncResult.syncedAt` change from `Date` to `string` on the API side, converted with `.toISOString()` at the point each service returns its result. Both were already serialized to an ISO string on the wire (`JSON.stringify` does this to any `Date`) and the UI's hand-written copies of these types already declared `syncedAt` as `string` — so this fixes a pre-existing type inaccuracy (the API's own type didn't match what it actually sent) rather than changing wire behavior.
+- Any new type or pure function needed by both packages should be checked against the identical-only rule before it goes into `packages/shared`: if the two packages' versions would have to differ in even one field's type, it belongs in each package's own tree, not here.
+- `pnpm fallow:baseline`, re-run after this move, no longer has to baseline the two cross-package clone groups this ADR resolves — they simply don't reproduce.
