+++ packages/api/eslint.config.mjs
@@ -18,4 +18,20 @@ export default antfu(
       'ts/no-explicit-any': 'error',
     },
   },
+  {
+    files: [
+      'src/**/*.entity.ts',
+      'src/migrations/**/*.ts',
+      'src/config/typeorm.config.ts',
+    ],
+    rules: {
+      'ts/no-restricted-imports': ['error', {
+        paths: [{
+          name: 'kuroshiro-shared',
+          allowTypeImports: true,
+          message: 'Entities and migrations are compiled unbundled; only `import type` from kuroshiro-shared here (ADR-0020).',
+        }],
+      }],
+    },
+  },
 )
