+++ packages/api/package.json
@@ -58,6 +58,7 @@
     "@types/node": "catalog:",
     "@vitest/coverage-v8": "catalog:",
     "eslint": "catalog:",
+    "kuroshiro-shared": "workspace:*",
     "ts-loader": "catalog:",
     "ts-node": "catalog:",
     "tsup": "catalog:",
