+++ packages/api/src/device-models/device-models.controller.ts
@@ -1,7 +1,8 @@
+import type { DeviceModelSyncResult } from 'kuroshiro-shared'
 import { Body, Controller, Delete, Get, Logger, Param, Post, ServiceUnavailableException, UsePipes, ValidationPipe } from '@nestjs/common'
 import { getErrorMessage } from '../utils/getErrorMessage'
 import { CustomPalettesService } from './custom-palettes.service'
-import { DeviceModelSyncResult, DeviceModelSyncService } from './device-model-sync.service'
+import { DeviceModelSyncService } from './device-model-sync.service'
 import { DeviceModelsService } from './device-models.service'
 import { CreateCustomPaletteDto } from './dto/create-custom-palette.dto'
 import { DeviceModel } from './entities/device-model.entity'
