+++ packages/api/src/maintenance/maintenance.service.ts
@@ -1,3 +1,4 @@
+import type { BrokenScreen, CleanupResult, MaintenanceIssues, OrphanedDeviceDir, OrphanedScreenFile, TempFile } from 'kuroshiro-shared'
 import * as fs from 'node:fs'
 import * as path from 'node:path'
 import { Injectable, Logger } from '@nestjs/common'
@@ -7,51 +8,6 @@ import { Device } from '../devices/devices.entity'
 import { Screen } from '../screens/screens.entity'
 import { getErrorMessage } from '../utils/getErrorMessage'
 import { resolveAppPath } from '../utils/pathHelper'
-
-export interface OrphanedScreenFile {
-  deviceId: string
-  screenId: string
-  path: string
-  size: number
-}
-
-export interface OrphanedDeviceDir {
-  deviceId: string
-  path: string
-  fileCount: number
-  size: number
-}
-
-export interface BrokenScreen {
-  screenId: string
-  deviceId: string
-  filename: string
-  type: string
-}
-
-export interface TempFile {
-  path: string
-  age: number
-  size: number
-}
-
-export interface MaintenanceIssues {
-  orphanedScreenFiles: OrphanedScreenFile[]
-  orphanedDeviceDirs: OrphanedDeviceDir[]
-  brokenScreens: BrokenScreen[]
-  tempFiles: TempFile[]
-  oldUploads: TempFile[]
-  totalSize: number
-  scannedAt: string
-}
-
-export interface CleanupResult {
-  filesDeleted: number
-  dirsDeleted: number
-  screensDeleted: number
-  bytesFreed: number
-  errors: string[]
-}
 
 interface DeviceDirEntry {
   deviceId: string
