+++ packages/shared/eslint.config.mjs
@@ -1,0 +1,10 @@
+import antfu from '@antfu/eslint-config'
+
+export default antfu(
+  {},
+  {
+    rules: {
+      'ts/no-explicit-any': 'error',
+    },
+  },
+)
